Split UserEventLog collection resources
